:wheel_of_dharma: Ensure that private media uploads can be served
Normally, we'd do this with nginx where the django app sends the
X-Accel-Redirect header. However, this requires the Django container
and nginx container to have access to the same persistent volume
holding the actual files, which is currently not possible.

The nginx config is correct, but commented out so that you can
kubectl apply -f ./infra/drc/ without issues.

The SENDFILE_BACKEND envvar in docker defaults to
sendfile.backends.nginx. We override it to let Django serve the
files instead (which WILL hammer our uwsgi workers). This is a
working, but non-ideal situation.