:bug: Fix typo so that annotations work
Yeah... I don't know either where my Brian was at.

Changing annotations in the template causes pods to be stopped and
replaced conforming to the 'new' spec. Combined with an 'Always'
ImagePullPolicy, this effectively updates to the latest version of
an image, even if it's pinned on something like latest/develop.