:pencil: Add Zaken endpoint filter parameters to support werkvoorraad
Building a "werkvoorraad" or collection of zaken initiated by a
particular "Natuurlijk Persoon" was not possible without serious
performance penalties.

The new filters support filtering zaken based on the "rol" that a
"betrokkene" has in each ZAAK. Special attention has been paid to the
"boundaries of the API landscape", because of a lack of (ready for
production) API's for medewerkers/natuurlijke personen.

Filters added:

* rol__betrokkeneType
* rol__betrokkene
* rol__omschrijvingGeneriek
* rol__betrokkeneIdentificatie__natuurlijkPersoon__inpBsn
* rol__betrokkeneIdentificatie__medewerker__identificatie