:alien: Ensure drc deployments are recreated
... rather than performing a rolling update. A rolling update requires volumes
to be mounted into multiple pods, which requires the volumes to be RWX.

We don't have RWX volumes (yet), so to deal with RWO volumes, pods
must be killed first before a new one can be started.

Note that this is a breaking change - you need to delete the old deployment
first, as K8S cannot change the strategy type of existing deployments.