Refactor: Remove obsolete Jekyll setup and unused _assets directory
This commit removes the following artifacts of the older, manual Ruby Jekyll setup:

- Gemfile
- Gemfile.lock
- old_config.yml
- Docker files related to manual documenation builds
- _assets directory (assets is a 100% copy of this directory)

These files are no longer needed as the project is migrating to Docusaurus for documentation generation. The `assets` directory has been checked to be the actively used asset directory.

This cleanup removes unnecessary files, paving the way for a transition to Docusaurus.